+++ app/src/main/java/com/eatssu/android/presentation/map/MapFragmentView.kt
@@ -1,13 +1,11 @@
 package com.eatssu.android.presentation.map
 
 import android.content.Intent
-import android.graphics.PointF
 import androidx.compose.foundation.layout.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
